requirements.txt
pytest-freezegun==0.3.0.post1
requests==2.22.0
Flask==1.0.2
gunicorn==19.9.0
jsonschema==3.0.2
sendgrid==6.0.5
opencv-python==4.1.0.25
numpy==1.17.0
flask-cors==3.0.8
langdetect==1.0.7
psycopg2-binary==2.8.3
flask-log-request-id
requests_futures==1.0.0
